--- a/core/app/models/spree/exchange.rb
+++ b/core/app/models/spree/exchange.rb
@@ -18,12 +18,23 @@ module Spree
     end
 
     def perform!
-      shipments = Spree::Stock::Coordinator.new(@order, @reimbursement_objects.map(&:build_exchange_inventory_unit)).shipments
-      if shipments.flat_map(&:inventory_units).size != @reimbursement_objects.size
+      new_exchange_inventory_units = @reimbursement_objects.map(&:build_exchange_inventory_unit)
+      shipments = Spree::Stock::Coordinator.new(@order, new_exchange_inventory_units).shipments
+      shipments_units = shipments.flat_map(&:inventory_units)
+
+      if shipments_units.sum(&:quantity) != new_exchange_inventory_units.sum(&:quantity)
         raise UnableToCreateShipments.new("Could not generate shipments for all items. Out of stock?")
       end
+
+      @reimbursement_objects.each do |reimbursement_object|
+        reimbursement_object.exchange_inventory_unit = shipments_units.detect do |unit|
+          unit.variant_id == reimbursement_object.exchange_variant_id
+        end
+      end
+
       @order.shipments += shipments
       @order.save!
+
       shipments.each do |shipment|
         shipment.update!(@order)
         shipment.finalize!