--- a/core/app/models/spree/exchange.rb
+++ b/core/app/models/spree/exchange.rb
@@ -18,18 +18,12 @@ module Spree
     end
 
     def perform!
-      new_exchange_inventory_units = @reimbursement_objects.map(&:build_exchange_inventory_unit)
+      new_exchange_inventory_units = @reimbursement_objects.map(&:build_default_exchange_inventory_unit)
       shipments = Spree::Stock::Coordinator.new(@order, new_exchange_inventory_units).shipments
       shipments_units = shipments.flat_map(&:inventory_units)
 
       if shipments_units.sum(&:quantity) != new_exchange_inventory_units.sum(&:quantity)
         raise UnableToCreateShipments.new("Could not generate shipments for all items. Out of stock?")
-      end
-
-      @reimbursement_objects.each do |reimbursement_object|
-        reimbursement_object.exchange_inventory_unit = shipments_units.detect do |unit|
-          unit.variant_id == reimbursement_object.exchange_variant_id
-        end
       end
 
       @order.shipments += shipments