@priyank-gupta 2017-03-07
@Mafi88 I am done with the suggested changes. Also rebased with latest master.